requirements.txt
flask==2.3.2
flask-sqlalchemy==3.0.3
flask-cors==3.0.10
python-dotenv==1.0.0
spotipy==2.23.0
sqlalchemy==2.0.19